refactor to have an internal TernaryTree
